requirements.txt
Django
djangorestframework
django-debug-toolbar
mysqlclient